Extract Scraped classes to lib/ directory
